Module tidb_query_common::execute_stats

Structs

ExecSummary
Execution summaries to support EXPLAIN ANALYZE statements. We don’t use ExecutorExecutionSummary directly since it is less efficient.

ExecSummaryCollectorDisabled
A ExecSummaryCollector that does not collect anything. Acts like collect = false.

ExecSummaryCollectorEnabled
A normal ExecSummaryCollector that simply collects execution summaries. It acts like collect = true.

ExecuteStats
Execution statistics to be flowed between parent and child executors at once during collect_exec_stats() invocation.

WithSummaryCollector
Combines an ExecSummaryCollector with another type. This inner type T typically Executor/BatchExecutor, such that WithSummaryCollector<C, T> would implement the same trait and collects the statistics into C.

Traits

ExecSummaryCollector
A trait for all execution summary collectors.